change viint.h macros

+++ include/ultra64/viint.h
@@ -17,12 +17,12 @@
 // For use in initializing OSViMode structures
 
 #define BURST(hsync_width, color_width, vsync_width, color_start) \
-    (hsync_width | (color_width << 8) | (vsync_width << 16) | (color_start << 20))
+    ((u8)(hsync_width) | ((u8)(color_width) << 8) | ((u8)(vsync_width) << 16) | ((u16)(color_start) << 20))
 #define WIDTH(v) v
 #define VSYNC(v) v
 #define HSYNC(duration, leap) (duration | (leap << 16))
-#define LEAP(upper, lower) ((upper << 16) | lower)
-#define START(start, end) ((start << 16) | end)
+#define LEAP(upper, lower) (((u16)(upper) << 16) | (u16)(lower))
+#define START(start, end) (((u16)(start) << 16) | (u16)(end))
 
 #define FTOFIX(val, i, f) ((u32)(val * (f32)(1 << f)) & ((1 << (i + f)) - 1))
